+++ homework/java/kaydunov/lecture4/task3_ColorDetector/LightColorDetector.java
@@ -1,4 +1,4 @@
-package task2_ColorDetector;
+package task3_ColorDetector;
 
 public class LightColorDetector {
     //эту задачу мне скучно делать
